+++ src/test/java/dinkplugin/notifiers/KillCountNotifierTest.java
@@ -712,6 +712,29 @@ class KillCountNotifierTest extends MockedNotifierTest {
         assertDoesNotThrow(() -> RuneLiteAPI.GSON.toJson(body));
     }
 
+    @Test
+    void testNotifyBryophytaChest() {
+        // update config mocks
+        when(config.killCountInterval()).thenReturn(1);
+
+        // fire events
+        String gameMessage = "Your Bryophyta chests opened count is: 5.";
+        notifier.onGameMessage(gameMessage);
+        notifier.onTick();
+
+        // check notification
+        verifyCreateMessage(
+            PRIMARY_WEBHOOK_URL,
+            true,
+            NotificationBody.builder()
+                .text(buildTemplate("Bryophyta chests", 5))
+                .extra(new BossNotificationData("Bryophyta chests", 5, gameMessage, null, null, null, null))
+                .playerName(PLAYER_NAME)
+                .type(NotificationType.KILL_COUNT)
+                .build()
+        );
+    }
+
     private static Template buildTemplate(String boss, int count) {
         return Template.builder()
             .template(PLAYER_NAME + " has defeated {{boss}} with a completion count of " + count)