+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
 
 - Minor: Add (Corrupted) Gauntlet as a configurable safe death exception. (#819)
+- Minor: Fire kill count notifications for Obor and Bryophyta chest openings. (#820)
+- Bugfix: Identify Grid Master worlds if world selector was not opened. (#822)
+- Dev: Designate Grid Master as a `seasonalWorld` in notification metadata. (#818)
 
 ## 1.11.15